// client/src/types/UpdateHandoutRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { HandoutVisibility } from "./HandoutVisibility";

export type UpdateHandoutRequest = { title: string | null, content: string | null, visibility: HandoutVisibility | null, player_ids: Array<string> | null, };